修正ap4

+++ backend/src/main/java/in/tech_camp/prototype_d/custom_user/CustomUserDetail.java
@@ -18,8 +18,8 @@ public class CustomUserDetail implements UserDetails {
         this.user = user;
     }
 
-    // 便利なヘルパーメソッド
-    public Integer getId() {
+    // 便利なヘルパーメソッド（戻り値を Integer から Long に変更）
+    public Long getId() {
         return user != null ? user.getId() : null;
     }
